Выносит форму авторизации в компонент
